Refactor and merge APIs for updating background and fetching taxa details
